Remove token_program.rs
