+++ programs/native/erc20/src/lib.rs
@@ -10,11 +10,32 @@ extern crate serde_derive;
 #[macro_use]
 extern crate solana_sdk;
 
-use solana_sdk::account::KeyedAccount;
+use solana_sdk::account::{Account, KeyedAccount};
+use solana_sdk::native_loader;
 use solana_sdk::pubkey::Pubkey;
 use std::sync::{Once, ONCE_INIT};
 
 mod token_program;
+
+const ERC20_NAME: &str = "solana_erc20";
+const ERC20_PROGRAM_ID: [u8; 32] = [
+    131, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+    0,
+];
+
+pub fn id() -> Pubkey {
+    Pubkey::new(&ERC20_PROGRAM_ID)
+}
+
+pub fn account() -> Account {
+    Account {
+        tokens: 1,
+        owner: id(),
+        userdata: ERC20_NAME.as_bytes().to_vec(),
+        executable: true,
+        loader: native_loader::id(),
+    }
+}
 
 solana_entrypoint!(entrypoint);
 fn entrypoint(
