+++ src/lib.rs
@@ -77,7 +77,6 @@ pub mod streamer;
 pub mod system_program;
 pub mod system_transaction;
 pub mod thin_client;
-pub mod token_program;
 pub mod tpu;
 pub mod tpu_forwarder;
 pub mod tvu;
@@ -127,6 +126,7 @@ extern crate solana_jsonrpc_core as jsonrpc_core;
 extern crate solana_jsonrpc_http_server as jsonrpc_http_server;
 #[macro_use]
 extern crate solana_jsonrpc_macros as jsonrpc_macros;
+extern crate solana_erc20;
 extern crate solana_jsonrpc_pubsub as jsonrpc_pubsub;
 extern crate solana_jsonrpc_ws_server as jsonrpc_ws_server;
 extern crate solana_metrics;
